@@ -46,7 +46,7 @@ type Logger interface {
 	// the log line.  The key/value pairs can then be used to add additional
 	// variable information.  The key/value pairs should alternate string
 	// keys and arbitrary values.
-	Info(msg string, keysAndValues ...interface{})
+	Info(msg string, keysAndValues ...any)
 
 	// Error logs an error, with the given message and key/value pairs as context.
 	// It functions similarly to calling Info with the "error" named value, but may
@@ -56,7 +56,7 @@ type Logger interface {
 	// The msg field should be used to add context to any underlying error,
 	// while the err field should be used to attach the actual error that
 	// triggered this log line, if present.
-	Error(err error, msg string, keysAndValues ...interface{})
+	Error(err error, msg string, keysAndValues ...any)
 
 	// V returns a Logger value for a specific verbosity level, relative to
 	// this Logger.  In other words, V values are additive.  V higher verbosity
@@ -66,7 +66,7 @@ type Logger interface {
 
 	// WithValues adds some key-value pairs of context to a logger.
 	// See Info for documentation on how key/value pairs work.
-	WithValues(keysAndValues ...interface{}) logr.Logger
+	WithValues(keysAndValues ...any) logr.Logger
 
 	// WithName adds a new element to the logger's name.
 	// Successive calls with WithName continue to append
