@@ -19,6 +19,8 @@ package v1beta1
 import (
 	"strings"
 
+	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
+
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/validation/field"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -88,7 +90,7 @@ func defaultContainerManager(spec *KKMachineSpec) {
 var _ webhook.Validator = &KKMachine{}
 
 // ValidateCreate implements webhook.Validator so a webhook will be registered for the type
-func (k *KKMachine) ValidateCreate() error {
+func (k *KKMachine) ValidateCreate() (admission.Warnings, error) {
 	kkmachinelog.Info("validate create", "name", k.Name)
 
 	var allErrs field.ErrorList
@@ -97,7 +99,7 @@ func (k *KKMachine) ValidateCreate() error {
 }
 
 // ValidateUpdate implements webhook.Validator so a webhook will be registered for the type
-func (k *KKMachine) ValidateUpdate(old runtime.Object) error {
+func (k *KKMachine) ValidateUpdate(old runtime.Object) (admission.Warnings, error) {
 	kkmachinelog.Info("validate update", "name", k.Name)
 	var allErrs field.ErrorList
 	allErrs = append(allErrs, validateRepository(k.Spec.Repository)...)
@@ -105,9 +107,9 @@ func (k *KKMachine) ValidateUpdate(old runtime.Object) error {
 }
 
 // ValidateDelete implements webhook.Validator so a webhook will be registered for the type
-func (k *KKMachine) ValidateDelete() error {
+func (k *KKMachine) ValidateDelete() (admission.Warnings, error) {
 	kkmachinelog.Info("validate delete", "name", k.Name)
-	return nil
+	return nil, nil
 }
 
 func validateRepository(repo *Repository) field.ErrorList { //nolint:unparam
