@@ -42,10 +42,10 @@ const (
 	DefaultEtcdVersion             = "v3.5.6"
 	DefaultEtcdPort                = "2379"
 	DefaultDockerVersion           = "24.0.6"
-	DefaultContainerdVersion       = "1.7.8"
-	DefaultRuncVersion             = "v1.1.10"
-	DefaultCrictlVersion           = "v1.24.0"
-	DefaultKubeVersion             = "v1.23.10"
+	DefaultContainerdVersion       = "1.7.12"
+	DefaultRuncVersion             = "v1.1.11"
+	DefaultCrictlVersion           = "v1.29.0"
+	DefaultKubeVersion             = "v1.23.15"
 	DefaultCalicoVersion           = "v3.26.1"
 	DefaultFlannelVersion          = "v0.21.3"
 	DefaultFlannelCniPluginVersion = "v1.1.2"
@@ -54,7 +54,7 @@ const (
 	DefaulthybridnetVersion        = "v0.8.6"
 	DefaultKubeovnVersion          = "v1.10.6"
 	DefalutMultusVersion           = "v3.8"
-	DefaultHelmVersion             = "v3.9.0"
+	DefaultHelmVersion             = "v3.13.3"
 	DefaultDockerComposeVersion    = "v2.2.2"
 	DefaultRegistryVersion         = "2"
 	DefaultHarborVersion           = "v2.5.3"
