@@ -28,8 +28,8 @@ import (
 	"k8s.io/utils/pointer"
 	"sigs.k8s.io/cluster-api/util"
 
-	"github.com/kubesphere/kubekey/v3/test/e2e/framework"
-	"github.com/kubesphere/kubekey/v3/test/e2e/framework/clusterctl"
+	"sigs.k8s.io/cluster-api/test/framework"
+	"sigs.k8s.io/cluster-api/test/framework/clusterctl"
 )
 
 // QuickStartSpecInput is the input for QuickStartSpec.
